# api/openapi-v1.yml
openapi: 3.0.3
info:
  title: Thesis Discovery Platform API
  version: 0.1.0
  license:
    name: Unlicensed
    url: https://choosealicense.com/no-permission/
  description: |
    Simplified API contract for the thesis discovery platform.

    The system has four services:

    - Thesis Service: owns PostgreSQL data and exposes the frontend API.
    - Scraping Service: periodically scrapes chair/source websites and sends extracted results to the Thesis Service.
    - GenAI Service: converts raw scraped content into structured thesis entries.
    - Vector Search Service: owns the vector database and supports semantic search plus reindexing.

    Important design decision:
    During every successful scrape run for a chair, the Thesis Service replaces all existing thesis proposals
    for that chair in one transaction before instructing the Vector Search Service to refresh the
    corresponding vector index entries.
security: []
servers:
  - url: http://localhost:8080
    description: Thesis Service
  - url: http://localhost:8081
    description: Scraping Service
  - url: http://localhost:8082
    description: Vector Search Service
  - url: http://localhost:8000
    description: GenAI Service
...
paths:
...
  /internal/v1/thesis-service/scrape-runs:
    post:
      tags:
        - Thesis Service Internal API
      summary: Log a scrape run result
      description: |
        Called by the Scraping Service to record the outcome of a scrape attempt.
        This should be called for both successful and failed runs to maintain a complete history.
      operationId: logScrapeRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScrapeRunLogRequest'
      responses:
        '201':
          description: Scrape run logged successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScrapeRunLogResponse'
  /internal/v1/thesis-service/source-endpoints/{sourceEndpointId}/theses:
    put:
      tags:
        - Thesis Service Internal API
      summary: Replace all thesis proposals for a chair
      description: |
        Called by the Scraping Service only after successful scraping and successful
        GenAI extraction for one chair/source endpoint.

        Thesis Service behavior:
        1. In one transaction, delete all existing thesis proposals for `chairId`
           and insert the newly extracted thesis proposals.
        2. Commit only if the full replacement succeeds.
        3. After relational replacement succeeds, call the Vector Search Service
           to replace the vector index entries for the same chair.

        If validation or insertion fails, the old thesis data must remain unchanged.
      operationId: replaceChairTheses
      parameters:
        - $ref: '#/components/parameters/SourceEndpointId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceEndpointThesesReplacementRequest'
            example:
              theses:
                - title: "Semantic Search for Thesis Discovery"
                  degreeType: "MASTER"
                  originalDescription: "This thesis investigates semantic search for university thesis proposals..."
                  aiOverview: "Build and evaluate a semantic search pipeline for thesis discovery."
                  researchArea: "Artificial Intelligence"
                  sourceUrl: "https://example-chair.tum.de/theses/semantic-search"
                  status: "OPEN"
                  advisors:
                    - name: "Max Mustermann"
                      email: "max.mustermann@example.com"
                      profileUrl: "https://example-chair.tum.de/team/max"
      responses:
        '200':
          description: Chair thesis data replaced successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceEndpointThesesReplacementResponse'
        '400':
          description: Invalid replacement payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /thesis-internal/v1/source-endpoints/{sourceEndpointId}/detect-changes:
    post:
      tags:
        - Thesis Service Internal API
      summary: Detect if source endpoint content has changed
      description: |
        Called by the Scraping Service before calling the GenAI Service.
        The Thesis Service normalizes the HTML, hashes it, and checks if it matches the stored hash.
      operationId: detectChanges
      parameters:
        - $ref: '#/components/parameters/SourceEndpointId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetectChangesRequest'
      responses:
        '200':
          description: Change status computed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectChangesResponse'
        '404':
          description: Source endpoint not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    SourceEndpointId:
      name: sourceEndpointId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Internal source endpoint id.
...
  schemas:
...
    ThesisProposalInput:
      type: object
      required:
        - title
        - sourceUrl
      properties:
        title:
          type: string
        degreeType:
          type: string
          nullable: true
        originalDescription:
          type: string
          nullable: true
        aiOverview:
          type: string
          nullable: true
        researchArea:
          type: string
          nullable: true
        sourceUrl:
          type: string
          format: uri
        status:
          type: string
          default: "OPEN"
        advisors:
          type: array
          items:
            $ref: '#/components/schemas/AdvisorInput'
...
    AdvisorInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        profileUrl:
          type: string
          format: uri
          nullable: true
    ScrapeRunLogRequest:
      type: object
      required:
        - sourceEndpointId
        - startedAt
        - finishedAt
        - status
      properties:
        sourceEndpointId:
          type: integer
          format: int64
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - SUCCESS
            - PARTIAL_SUCCESS
            - FAILED
        errorMessage:
          type: string
          nullable: true
        rawHtmlSnapshot:
          type: string
          nullable: true
          description: Raw HTML snapshot of the source page, stored for debugging.
        candidatesFound:
          type: integer
          nullable: true
          example: 5
    ScrapeRunLogResponse:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
    SourceEndpointThesesReplacementRequest:
      type: object
      required:
        - theses
      properties:
        theses:
          type: array
          items:
            $ref: '#/components/schemas/ThesisProposalInput'
        lastContentHash:
          type: string
          nullable: true
          example: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
    SourceEndpointThesesReplacementResponse:
      type: object
      required:
        - sourceEndpointId
        - deletedRelationalTheses
        - insertedRelationalTheses
        - replacedVectorEntries
      properties:
        sourceEndpointId:
          type: integer
          format: int64
          example: 10
        deletedRelationalTheses:
          type: integer
          example: 12
        insertedRelationalTheses:
          type: integer
          example: 9
        replacedVectorEntries:
          type: integer
          example: 9
        errorMessage:
          type: string
          nullable: true
          example: "Vector search sync failed: connection refused"
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true
    DetectChangesRequest:
      type: object
      required:
        - rawHtml
      properties:
        rawHtml:
          type: string
          description: Raw HTML of the scraped page to be normalized and checked.
    DetectChangesResponse:
      type: object
      required:
        - changed
        - contentHash
      properties:
        changed:
          type: boolean
          description: True if the normalized content hash differs from the last successfully processed hash.
        sanitizedHtml:
          type: string
          nullable: true
          description: Cleaned semantic HTML that is safe and optimized for sending to GenAI.
        contentHash:
          type: string
          description: The SHA-256 hash of the normalized text content.